fixed bugs, now properly display recipe page, resolved conflict
