+++ source/recipe-template.html
@@ -14,7 +14,7 @@
 
     <!-- Main Stylesheets & Scripts -->
     <link rel="stylesheet" href="./styles/recipe-template.css" />
-    <script src="./scripts/script.js"></script>
+    <script src="./scripts/script.js" type="module"></script>
     <!-- Category Card Custom Element -->
     <script src="./scripts/CategoryCard.js" type="module"></script>
     <!-- Recipe Card Custom Element -->
