+++ source/scripts/RecipePage.js
@@ -11,116 +11,115 @@ class RecipePage extends HTMLElement{
       style.innerHTML = `
       .header{
         text-align: center;
-    }
-    
-    .header h1{
-        position: relative;
-        display: inline;
-        font-style: normal;
-        font-weight: normal;
-        font-size: 64px;
-        line-height: 96px;
-        text-align: center;
-    }
-  
-    .header #bookmark{
-      text-align: center;
-      /* padding-left: 50px; */
-      margin-left : 50px;
-      /*float: right;*/
-    }
-  
-    .middle > div > h3{
-        text-align: center;
-        font-weight: bold;
-        margin-top: 80px;
-    }
-  
-    #clear-checkboxes{
-      position: absolute;
-      top: 50%;
-      left: 50%;
-      transform: translate(-50%, -50%);
-      margin-top: 40px;
-    }
-  
-    /*-- custom checkbox style --*/
-    
-    input[class="ingredients-custom-checkbox"] + label,
-    input[class="ingredients-custom-checkbox"] + label::before
-    {
-        display: inline-block;
-        vertical-align: middle;
-        
-    }
-  
-    input[class="ingredients-custom-checkbox"]
-    {
-        opacity: 0; /*hide the actual checkbox*/
-        position: absolute;
-        transform: scale(2);
-    }
-  
-    input[class="ingredients-custom-checkbox"]:hover
-    {
-      cursor: pointer;
-    }
+      }
       
-    input[class="ingredients-custom-checkbox"] + label::before
-    {
-      /*custom new checkbox*/
-        content: "";
-        width: 15px;
-        height: 15px;
-        margin: 0 4px 0 0;
-        line-height: 14px;
+      .header h1{
+          position: relative;
+          display: inline;
+          font-style: normal;
+          font-weight: normal;
+          font-size: 64px;
+          line-height: 96px;
+          text-align: center;
+      }
+    
+      .header #bookmark{
+        cursor: pointer;
         text-align: center;
-        border: solid 3px var(--primary);
-    }
+        /* padding-left: 50px; */
+        margin-left : 50px;
+        /*float: right;*/
+      }
     
-      input[class="ingredients-custom-checkbox"]:checked + label::before
-    {
-        content: "\\2714"; /* tick sign */
-        font-size: 15px;
-    }
-  
-    #ingredients-list > ul > ol > label{
-      color: black;
-    }
-    /*-----------------------------------*/
+      .middle > div > h3{
+          text-align: center;
+          font-weight: bold;
+          margin-top: 80px;
+      }
     
-    #instructions > ol > li{
-      color: black;
-    }
-  
-    #ingredients-list{
-        /* text-align: center; */
-        font-style: normal;
-        font-weight: normal;
-  
-    }
-  
-    #ingredients-list > button{
-      border: 0;
-      background-color: inherit;
-      font: inherit;
-      text-align: center;
-      padding: 20px;
-    }
-  
-    #ingredients-list> button:hover{
-      cursor: pointer;
-      background: #eee;
-    }
+      #clear-checkboxes{
+        position: absolute;
+        left: 50%;
+        transform: translate(-50%, -50%);
+      }
+    
+      /*-- custom checkbox style --*/
+      
+      input[class="ingredients-custom-checkbox"] + label,
+      input[class="ingredients-custom-checkbox"] + label::before
+      {
+          display: inline-block;
+          vertical-align: middle;
+          
+      }
+    
+      input[class="ingredients-custom-checkbox"]
+      {
+          opacity: 0; /*hide the actual checkbox*/
+          position: absolute;
+          transform: scale(2);
+      }
+    
+      input[class="ingredients-custom-checkbox"]:hover
+      {
+        cursor: pointer;
+      }
+        
+      input[class="ingredients-custom-checkbox"] + label::before
+      {
+        /*custom new checkbox*/
+          content: "";
+          width: 15px;
+          height: 15px;
+          margin: 0 4px 0 0;
+          line-height: 14px;
+          text-align: center;
+          border: solid 3px var(--primary);
+      }
+      
+        input[class="ingredients-custom-checkbox"]:checked + label::before
+      {
+          content: "\\2714"; /* tick sign */
+          font-size: 15px;
+      }
+    
+      #ingredients-list > ul > ol > label{
+        color: black;
+      }
+      /*-----------------------------------*/
+      
+      #instructions > ol > li{
+        color: black;
+      }
+    
+      #ingredients-list{
+          /* text-align: center; */
+          font-style: normal;
+          font-weight: normal;
+    
+      }
+    
+      #ingredients-list > button{
+        border: 0;
+        background-color: inherit;
+        font: inherit;
+        text-align: center;
+        padding: 20px;
+      }
+    
+      #ingredients-list> button:hover{
+        cursor: pointer;
+        background: #eee;
+      }
       `;
 
       container.innerHTML = `
       <header class="header">
-      <h1></h1>
-      <input id="bookmark" onclick="setBookMark()" type="image" src="img/icons/bookmark-empty.svg" name="bookmark-empty" width="56" height="56">
+        <h1></h1>
+        <img id="bookmark" onclick="setBookMark()" src="img/icons/bookmark-empty.svg" name="bookmark-empty" width="56" height="56">
       </header>
-
       <main class="middle">
+        <img style="display: block; margin-left: auto; margin-right: auto;">
           <div id="ingredients-list">
               <h3>INGREDIENTS</h3>
               <ul style="list-style-type: none;">
@@ -144,14 +143,15 @@ class RecipePage extends HTMLElement{
             console.log("Error: no data exists");
             return;
         }
+        this.json = data;
 
       this.shadowRoot.querySelector("article").innerHTML = `
       <header class="header">
-          <h1></h1>
-          <input id="bookmark" onclick="setBookMark()" type="image" src="img/icons/bookmark-empty.svg" name="bookmark-empty" width="56" height="56">
+        <h1></h1>
+        <img id="bookmark" onclick="setBookMark()" src="./img/icons/bookmark-empty.svg" name="bookmark-empty" width="56" height="56">
       </header>
-
       <main class="middle">
+        <img style="display: block; margin-left: auto; margin-right: auto;" >
           <div id="ingredients-list">
               <h3>INGREDIENTS</h3>
               <ul style="list-style-type: none;">
@@ -167,7 +167,7 @@ class RecipePage extends HTMLElement{
       `;
 
         // Header - title
-    
+        this.shadowRoot.querySelector(".middle > img").src= data["image"];
         this.shadowRoot.querySelector(".header > h1").innerHTML = data["title"];
 
         //get ingredient list
@@ -191,6 +191,9 @@ class RecipePage extends HTMLElement{
             li.innerHTML = element;
             this.shadowRoot.querySelector("#instructions > ol").appendChild(li);
         });
+    }
+    get data(){
+      return this.json;
     }
 }
