+++ source/scripts/main.js
@@ -1,0 +1,167 @@
+var $SOMenuVisibility = "hidden";
+
+function toggleMenu() {
+    var menuIcon = document.getElementById("menu-icon");
+    menuIcon.classList.toggle("change");
+
+    var slideOverMenu = document.getElementById("slide-over-menu");
+
+    if ($SOMenuVisibility == "hidden") {
+        slideOverMenu.style.transform = "translate(100%)";
+        $SOMenuVisibility = "visible";
+    }
+    else {
+        slideOverMenu.style.transform = "translate(-100%)";
+        $SOMenuVisibility = "hidden";
+    }
+}
+
+function showSettings() {
+    hideHome();
+    hideCookbooks();
+    hideRecipeCards();
+    hideRecipePage();
+    const settings = document.getElementById("settings-container");
+    settings.style.visibility = "visible";
+    //settings.style.transform = "translate(100%)";
+    // Get the list of restrictions from local storage
+    const getDietaryRestrictions = JSON.parse(localStorage.getItem("dietaryRestrictions"));
+    const dietaryContainerElements = document.getElementById('dietary-container').elements;
+
+    for (let i = 0; i < dietaryContainerElements.length; i++) {
+        const dietaryRestriction = dietaryContainerElements[i];
+        // If our restriction is in the list, then check it on the page
+        if (getDietaryRestrictions.includes(dietaryRestriction.value)) {
+            dietaryRestriction.checked = true;
+        }
+    }
+}
+
+function clearCheckBoxes(){
+    let checkboxes = document.querySelector("#recipe-page-container > recipe-page").shadowRoot.querySelectorAll("#ingredients-list > ul > ol > input");
+    console.log(checkboxes);
+    checkboxes.forEach(e => e.checked = false);
+}
+
+function hideRecipePage(){
+    const recipePage = document.getElementById("recipe-page-container");
+    recipePage.classList.add("hidden");
+}
+
+function hideSettings() {
+    const settings = document.getElementById("settings-container");
+    settings.style.visibility = "hidden";
+    // settings.style.transform = "translate(-100%)";
+}
+
+function showHome() {
+    hideSettings();
+    hideCookbooks();
+    hideRecipeCards();
+    showCategoryCards();
+    hideRecipePage();
+    document.getElementById('search-query').value = ''; //clears search result
+    const search = document.getElementById("search");
+    search.style.visibility = "visible";
+}
+
+function hideHome() {
+    hideCategoryCards();
+    const search = document.getElementById("search");
+    search.style.visibility = "hidden";
+}
+
+function showCookbooks() {
+    hideSettings();
+    hideHome();
+    hideRecipeCards();
+    const cookbook = document.getElementById("cookbook-container");
+    cookbook.style.visibility = "visible";
+}
+
+function hideCookbooks() {
+    const cookbook = document.getElementById("cookbook-container");
+    cookbook.style.visibility = "hidden";
+}
+
+function showRecipePage(){
+    const recipePage = document.getElementById("recipe-page-container");
+    recipePage.classList.remove("hidden");
+}
+
+function showRecipeCards() {
+    const recipeCards = document.getElementById("recipe-card-container");
+    recipeCards.style.visibility = "visible";
+}
+function hideRecipeCards() {
+    const recipeCards = document.getElementById("recipe-card-container");
+    recipeCards.style.visibility = "hidden";
+}
+
+function showCategoryCards() {
+    const categoryCards = document.getElementById("category-wrapper");
+    categoryCards.style.visibility = "visible";
+    categoryCards.classList.remove("hidden");
+}
+
+function hideCategoryCards() {
+    const categoryCards = document.getElementById("category-wrapper");
+    categoryCards.style.visibility = "hidden";
+}
+
+function updateSettings() {
+    const dietaryRestrictionList = [];
+    // Get all the inputs under the div
+    const dietaryContainerElements = document.getElementById('dietary-container').elements;
+    for (let i = 0; i < dietaryContainerElements.length; i++) {
+        // If a checkbox is checked, then add it to our list
+        const inputElement = dietaryContainerElements[i];
+        if (inputElement.checked) {
+            dietaryRestrictionList.push(inputElement.value);
+        }
+    }
+
+    // Add list to local storage
+    localStorage.setItem("dietaryRestrictions", JSON.stringify(dietaryRestrictionList));
+
+    // TODO: add confirmation message in HTML (alert is temporary)
+    alert("your preferences have been updated");
+}
+
+function checkBookMark(data){
+    let bookmarkList = JSON.parse(localStorage.getItem("bookmark"));
+    const title = data["title"];
+    if (bookmarkList == null)
+        return;
+    if (bookmarkList[title] != null)
+    {
+        let bookMark = document.querySelector("#recipe-page-container > recipe-page").shadowRoot.querySelector("#bookmark");
+        bookMark.src = "./img/icons/bookmark-filled.svg";
+        bookMark.setAttribute("name", "bookmark-filled");
+    }
+}
+
+function setBookMark(){
+    // check local storage for bookmark
+    let bookmarkList = JSON.parse(localStorage.getItem("bookmark"));
+    if (bookmarkList == null)
+        bookmarkList = {};
+
+    let bookMark = document.querySelector("#recipe-page-container > recipe-page").shadowRoot.querySelector("#bookmark");
+    const name = bookMark.getAttribute("name");
+    const ID = document.querySelector("recipe-page").data["id"];
+    const title = document.querySelector("recipe-page").data["title"];
+    if (name == "bookmark-empty")
+    {
+        bookmarkList[title] = ID;
+        bookMark.src = "./img/icons/bookmark-filled.svg";
+        bookMark.setAttribute("name", "bookmark-filled");
+    }
+    else
+    {
+        delete bookmarkList[title];
+        bookMark.src = "./img/icons/bookmark-empty.svg";
+        bookMark.setAttribute("name", "bookmark-empty");
+    }
+    localStorage.setItem("bookmark", JSON.stringify(bookmarkList));
+}
